requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.txt -o requirements.txt
asgiref==3.8.1
    # via
    #   -r requirements.txt
    #   django
boto3==1.34.156
    # via -r requirements.txt
botocore==1.34.157
    # via
    #   -r requirements.txt
    #   boto3
    #   s3transfer
django==5.0.7
    # via
    #   -r requirements.txt
    #   django-extensions
    #   django-storages
django-extensions==3.2.3
    # via -r requirements.txt
django-storages==1.14.4
    # via -r requirements.txt
factory-boy==3.3.0
    # via -r requirements.txt
faker==26.3.0
    # via factory-boy
jmespath==1.0.1
    # via
    #   -r requirements.txt
    #   boto3
    #   botocore
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.txt
    #   botocore
    #   faker
s3transfer==0.10.2
    # via
    #   -r requirements.txt
    #   boto3
six==1.16.0
    # via
    #   -r requirements.txt
    #   python-dateutil
sqlparse==0.5.1
    # via
    #   -r requirements.txt
    #   django
urllib3==2.2.2
    # via
    #   -r requirements.txt
    #   botocore
